# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "AFSE API",
    "description": "API for Managing AFSE",
    "contact": {
      "name": "API Support",
      "email": "matteo.martinelli5@studenti.unimi.it"
    }
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/marvel/characters/{name}": {
      "get": {
        "tags": [
          "Marvel Service"
        ],
        "summary": "Recupera i personaggi Marvel che iniziano con un determinato nome",
        "description": "Recupera i personaggi Marvel il cui nome inizia con il valore specificato nel parametro \"name\". Richiede un JWT token per autenticare la richiesta.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Nome o parte iniziale del nome del personaggio da cercare",
            "example": "Spider"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista di personaggi recuperata con successo",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "array"
                },
                "items": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "object"
                    },
                    "properties": {
                      "type": "object",
                      "properties": {
                        "characterId": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "string"
                            },
                            "description": {
                              "type": "string",
                              "example": "ID del personaggio"
                            },
                            "example": {
                              "type": "string",
                              "example": "1009610"
                            }
                          }
                        },
                        "name": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "string"
                            },
                            "description": {
                              "type": "string",
                              "example": "Nome del personaggio"
                            },
                            "example": {
                              "type": "string",
                              "example": "Spider-Man"
                            }
                          }
                        },
                        "description": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "string"
                            },
                            "description": {
                              "type": "string",
                              "example": "Breve descrizione del personaggio"
                            },
                            "example": {
                              "type": "string",
                              "example": "Bitten by a radioactive spider, Peter Parker now has spider-like abilities."
                            }
                          }
                        },
                        "imageUrl": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "string"
                            },
                            "description": {
                              "type": "string",
                              "example": "URL dell'immagine del personaggio"
                            },
                            "example": {
                              "type": "string",
                              "example": "http://example.com/spider-man.jpg"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Nessun personaggio trovato con il nome specificato",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Messaggio di errore"
                        },
                        "example": {
                          "type": "string",
                          "example": "Nessun personaggio trovato"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Errore interno del server",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Messaggio di errore"
                        },
                        "example": {
                          "type": "string",
                          "example": "Internal server error"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "jwt",
        "in": "cookie",
        "name": "jwt",
        "description": "JWT token passato tramite cookie"
      }
    },
...
  }
}